requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
openai
pydantic
python-dotenv
langdetect